backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
pymysql==1.1.1
cryptography==42.0.5
pydantic==2.6.4
pydantic-core==2.16.3
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
celery==5.3.6
redis==5.0.3
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.1.0
dnspython==2.6.1
apscheduler==3.10.4
slowapi==0.1.9
email-validator==2.1.1
